requirements.txt
pandas~=2.2
scikit-learn~=1.5
jupyterlab~=4.2
joblib~=1.4
pyarrow~=16.0
python-dotenv
psycopg2-binary